Add staff member unique email validation
