+++ src/etools/applications/partners/tests/test_views.py
@@ -400,356 +400,6 @@ class TestPartnerOrganizationCreateView(BaseTenantTestCase):
         return response_json['id']
 
 
-class TestPartnerOrganizationRetrieveUpdateDeleteViews(BaseTenantTestCase):
-    '''Exercise the retrieve, update, and delete views for PartnerOrganization'''
-    @classmethod
-    def setUpTestData(cls):
-        cls.unicef_staff = UserFactory(is_staff=True)
-        cls.partner = PartnerFactory(
-            partner_type=PartnerType.CIVIL_SOCIETY_ORGANIZATION,
-            cso_type="International",
-            hidden=False,
-            vendor_number="DDD",
-            short_name="Short name",
-        )
-
-        report = "report.pdf"
-        cls.assessment1 = Assessment.objects.create(
-            partner=cls.partner,
-            type="Micro Assessment"
-        )
-        cls.assessment2 = Assessment.objects.create(
-            partner=cls.partner,
-            type="Micro Assessment",
-            report=report,
-            completed_date=datetime.date.today()
-        )
-
-        cls.partner_gov = PartnerFactory(partner_type=PartnerType.GOVERNMENT)
-
-        agreement = AgreementFactory(
-            partner=cls.partner,
-            signed_by_unicef_date=datetime.date.today())
-
-        cls.intervention = InterventionFactory(agreement=agreement)
-        cls.output_res_type = ResultTypeFactory(name=ResultType.OUTPUT)
-
-        cls.result = ResultFactory(
-            result_type=cls.output_res_type,)
-
-        cls.partnership_budget = InterventionBudget.objects.create(
-            intervention=cls.intervention,
-            unicef_cash=100,
-            unicef_cash_local=10,
-            partner_contribution=200,
-            partner_contribution_local=20,
-            in_kind_amount_local=10,
-        )
-        cls.amendment = InterventionAmendment.objects.create(
-            intervention=cls.intervention,
-            types=[InterventionAmendment.RESULTS]
-        )
-
-        cls.cp = CountryProgrammeFactory(__sequence=10)
-        cls.cp_output = ResultFactory(result_type=cls.output_res_type)
-
-    def test_api_partners_update_with_members(self):
-        self.assertFalse(Activity.objects.exists())
-        response = self.forced_auth_req(
-            'get',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data["staff_members"]), 1)
-        self.assertEqual(response.data["staff_members"][0]["first_name"], "Mace")
-
-        staff_members = [{
-            "title": "Some title",
-            "first_name": "John",
-            "last_name": "Doe",
-            "email": "a@a.com",
-            "active": True,
-        }]
-        data = {
-            "name": self.partner.name + ' Updated',
-            "partner_type": self.partner.partner_type,
-            "vendor_number": self.partner.vendor_number,
-            "staff_members": staff_members,
-        }
-        response = self.forced_auth_req(
-            'patch',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-            data=data,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data["staff_members"]), 2)
-
-        self.assertEqual(
-            Activity.objects.filter(action=Activity.UPDATE).count(),
-            1
-        )
-
-    def test_api_partners_update_invalid_basis_for_type_of_assessment(self):
-        data = {
-            "type_of_assessment": PartnerOrganization.HIGH_RISK_ASSUMED,
-            "basis_for_risk_rating": "NOT NULL VALUE",
-        }
-        response = self.forced_auth_req(
-            'patch',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-            data=data
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        error = [
-            ErrorDetail(string='The basis for risk rating has to be blank if Type is Low or High', code='invalid')]
-        self.assertEqual(response.data, {
-            "basis_for_risk_rating": error})
-
-    def test_api_partners_update_assessments_invalid(self):
-        self.assertFalse(Activity.objects.exists())
-        today = datetime.date.today()
-        assessments = [{
-            "id": self.assessment2.id,
-            "completed_date": datetime.date(today.year + 1, 1, 1),
-        }]
-        data = {
-            "assessments": assessments,
-        }
-        response = self.forced_auth_req(
-            'patch',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-            data=data,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(response.data, {"assessments":
-                                         {"completed_date": ["The Date of Report cannot be in the future"]}})
-        self.assertEqual(
-            Activity.objects.filter(action=Activity.UPDATE).count(),
-            0
-        )
-
-    def test_api_partners_update_assessments_longago(self):
-        self.assertFalse(Activity.objects.exists())
-        today = datetime.date.today()
-        assessments = [{
-            "id": self.assessment2.id,
-            "completed_date": datetime.date(today.year - 3, 1, 1),
-        }]
-        data = {
-            "assessments": assessments,
-        }
-        response = self.forced_auth_req(
-            'patch',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-            data=data,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            Activity.objects.filter(action=Activity.UPDATE).count(),
-            1
-        )
-
-    def test_api_partners_update_assessments_today(self):
-        self.assertFalse(Activity.objects.exists())
-        completed_date = datetime.date.today()
-        assessments = [{
-            "id": self.assessment2.id,
-            "completed_date": completed_date,
-        }]
-        data = {
-            "assessments": assessments,
-        }
-        response = self.forced_auth_req(
-            'patch',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-            data=data,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            Activity.objects.filter(action=Activity.UPDATE).count(),
-            1
-        )
-
-    def test_api_partners_update_assessments_yesterday(self):
-        self.assertFalse(Activity.objects.exists())
-        completed_date = datetime.date.today() - datetime.timedelta(days=1)
-        assessments = [{
-            "id": self.assessment2.id,
-            "completed_date": completed_date,
-        }]
-        data = {
-            "assessments": assessments,
-        }
-        response = self.forced_auth_req(
-            'patch',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-            data=data,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(
-            Activity.objects.filter(action=Activity.UPDATE).count(),
-            1
-        )
-
-    def test_api_partners_update_with_members_empty_phone(self):
-        self.assertFalse(Activity.objects.exists())
-        response = self.forced_auth_req(
-            'get',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(len(response.data["staff_members"]), 1)
-        self.assertEqual(response.data["staff_members"][0]["first_name"], "Mace")
-
-        staff_members = [{
-            "title": "Some title",
-            "first_name": "John",
-            "last_name": "Doe",
-            "email": "a1@a.com",
-            "active": True,
-            "phone": ''
-        }]
-        data = {
-            "staff_members": staff_members,
-        }
-        response = self.forced_auth_req(
-            'patch',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-            data=data,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["staff_members"][1]["phone"], '')
-        self.assertEqual(
-            Activity.objects.filter(action=Activity.UPDATE).count(),
-            1
-        )
-
-    def test_api_partners_update_assessments_tomorrow(self):
-        self.assertFalse(Activity.objects.exists())
-        completed_date = datetime.date.today() + datetime.timedelta(days=1)
-        assessments = [{
-            "id": self.assessment2.id,
-            "completed_date": completed_date,
-        }]
-        data = {
-            "assessments": assessments,
-        }
-        response = self.forced_auth_req(
-            'patch',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-            data=data,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
-        self.assertEqual(response.data, {"assessments":
-                                         {"completed_date": ["The Date of Report cannot be in the future"]}})
-        self.assertEqual(
-            Activity.objects.filter(action=Activity.UPDATE).count(),
-            0
-        )
-
-    def test_api_partners_retrieve(self):
-        response = self.forced_auth_req(
-            'get',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertIn("vendor_number", response.data.keys())
-        self.assertIn("address", response.data.keys())
-        self.assertIn("Partner", response.data["name"])
-        self.assertEqual(['audits', 'programme_visits', 'spot_checks'],
-                         sorted(response.data['hact_min_requirements'].keys()))
-        self.assertEqual(['assurance_coverage', 'audits', 'outstanding_findings', 'programmatic_visits', 'spot_checks'],
-                         sorted(response.data['hact_values'].keys()))
-        self.assertCountEqual(
-            ['completed', 'minimum_requirements'],
-            response.data['hact_values']['audits'].keys())
-        self.assertEqual(response.data['interventions'], [])
-
-    def test_api_partners_retreive_actual_fr_amounts(self):
-        self.intervention.status = Intervention.ACTIVE
-        self.intervention.save()
-        fr_header_1 = FundsReservationHeaderFactory(intervention=self.intervention)
-        fr_header_2 = FundsReservationHeaderFactory(intervention=self.intervention)
-
-        response = self.forced_auth_req(
-            'get',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(Decimal(response.data["interventions"][0]["actual_amount"]),
-                         Decimal(fr_header_1.actual_amt_local + fr_header_2.actual_amt_local))
-
-    def test_api_partners_retrieve_staff_members(self):
-        response = self.forced_auth_req(
-            'get',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertIn("staff_members", response.data.keys())
-        self.assertEqual(len(response.data["staff_members"]), 1)
-
-    def test_api_partners_update(self):
-        self.assertFalse(Activity.objects.exists())
-        data = {
-            "name": "Updated name again",
-        }
-        response = self.forced_auth_req(
-            'patch',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-            data=data,
-        )
-
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertIn("Updated", response.data["name"])
-        self.assertEqual(
-            Activity.objects.filter(action=Activity.UPDATE).count(),
-            1
-        )
-
-    def test_api_partners_update_hidden(self):
-        # make some other type to filter against
-        self.assertFalse(Activity.objects.exists())
-        data = {
-            "hidden": True
-        }
-        response = self.forced_auth_req(
-            'patch',
-            reverse('partners_api:partner-detail', args=[self.partner.pk]),
-            user=self.unicef_staff,
-            data=data
-        )
-        self.assertEqual(response.status_code, status.HTTP_200_OK)
-        self.assertEqual(response.data["hidden"], False)
-        self.assertEqual(
-            Activity.objects.filter(action=Activity.UPDATE).count(),
-            1
-        )
-
-
 class TestPartnershipViews(BaseTenantTestCase):
     @classmethod
     def setUpTestData(cls):
